src/backend/requirements.txt
django-cors-headers>=2.4.0 
Pillow >= 7.0.0
django-multiselectfield
boto3
django-storages

